Updates Gotify to version 2.6.3
Updates the Gotify add-on to the latest version (2.6.3).

This change includes a new Dockerfile that uses the official Gotify image, simplifies the build process, and ensures compatibility with various architectures.
It also replaces the previous `run.sh` script with a new one that leverages `bashio` for configuration and logging, improving maintainability and user experience.